switch the doc of icon to doc.rs
